refactor: api 파일 모듈화 구조 분리
